Refine code.
This aims to get ready for the syntax analyze.
